requirements.txt
g4f==0.5.7.1
markitdown==0.1.2
uvicorn==0.35.0
fastapi==0.116.1
python-multipart==0.0.20
a2wsgi==1.10.10
Flask==3.1.1
discord.py==2.5.2
httpx==0.28.1